Wahab dev.

+++ app/Http/Controllers/ListingController.php
@@ -6,6 +6,7 @@ use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 
+
 class ListingController extends Controller
 {
 
@@ -42,6 +43,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public'); // 'logos' is the sub-folder and 'public' is the main folder
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing successfully created...!');
@@ -55,7 +58,11 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
-        // dd($request->file('logo'));
+        // Make sure the current loggedin user is authorized
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title'    => 'required',
             'company'  => ['required'],
@@ -78,7 +85,17 @@ class ListingController extends Controller
     // Deleting a post
     public function destroy(Listing $listing)
     {
+        // Make sure the current loggedin user is authorized
+        if ($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized Action");
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Post deleted successfully');
+    }
+
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
